+++ lib/models/user.js
@@ -27,6 +27,9 @@ module.exports = function (sequelize, DataTypes) {
       type: DataTypes.STRING,
       unique: true
     },
+    name: {
+      type: DataTypes.STRING
+    },
     profile: {
       type: DataTypes.TEXT
     },
@@ -68,12 +71,14 @@ module.exports = function (sequelize, DataTypes) {
       constraints: false
     })
   }
+
   User.getProfile = function (user) {
     if (!user) {
       return null
     }
-    return user.profile ? User.parseProfile(user.profile) : (user.email ? User.parseProfileByEmail(user.email) : null)
+    return user.profile ? User.parseProfile(user.profile) : (user.email ? User.parseProfileByUser(user) : null)
   }
+
   User.parseProfile = function (profile) {
     try {
       profile = JSON.parse(profile)
@@ -90,6 +95,7 @@ module.exports = function (sequelize, DataTypes) {
     }
     return profile
   }
+
   User.parsePhotoByProfile = function (profile, bigger) {
     var photo = null
     switch (profile.provider) {
@@ -146,9 +152,11 @@ module.exports = function (sequelize, DataTypes) {
     }
     return photo
   }
-  User.parseProfileByEmail = function (email) {
+
+  User.parseProfileByUser = function (user) {
+    var email = user.email
     return {
-      name: email.substring(0, email.lastIndexOf('@')),
+      name: user.name || email.substring(0, email.lastIndexOf('@')),
       photo: generateAvatarURL('', email, false),
       biggerphoto: generateAvatarURL('', email, true)
     }
